Move STM Unique ID stuff out into init.c
